pyproject.toml
[project]
name = "fastapi-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cloudinary>=1.44.1",
    "fastapi>=0.116.1",
    "openai-agents>=0.2.4",
    "openai-whisper>=20250625",
    "pymongo>=4.13.2",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "pyttsx3>=2.99",
    "rapidfuzz>=3.13.0",
    "requests>=2.32.4",
    "uvicorn>=0.35.0",
]
